requirements.txt
altair
pandas
streamlit
sentence-transformers
faiss-cpu>=1.7.4
PyPDF2>=3.0.1
python-dotenv>=1.0.0
transformers>=4.30.0
numpy>=1.24.0
tqdm>=4.65.0
scikit-learn>=1.3.0
huggingface-hub>=0.16.0
torch
accelerate